Lab3+4: Backend/backend/requirements.txt
fastapi==0.118.2
pillow==11.3.0
